skills/video-translation/scripts/base/pyproject.toml
[project]
name = "translate-video-base-scripts"
version = "0.1.0"
description = "Utility scripts for translate-video"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "demucs",
     "qwen-asr",
     "torch",
     "moviepy",
     "ffmpeg-python",
     "soundfile",
     "pyyaml",
     "python-dotenv",
     "requests",
     "librosa",
     "nagisa==0.2.11",
     "soynlp==0.0.493",
     "accelerate==1.12.0",
     "qwen-omni-utils",
     "torchcodec",
     "soxr>=1.0.0",

]

[project.optional-dependencies]
vllm = [
    "vllm==0.14.0",
]

[tool.uv]
dev-dependencies = [
    "pytest",
    "black",
    "isort"
]

